requirements.txt
numpy~=1.24.3
yaml~=0.2.5
pyyaml~=6.0.1
datasets~=2.12.0
pandas~=1.5.2
scikit-learn~=1.2.2
torchvision~=0.13.1
pillow~=10.0.1